feat: UNT-T20403: Work on Audit Changes - Fix typo in readme - yAxisMax description updated in readme - added keywords to package json - removed contributors from package json - updated package json description: Personalized graphs featuring customizable options for React Native app - updated package json version to 1.0.0 - fixed typo Commiting in Contributing MD file - renamed ToolTip file names, because it is a single word - updated BarChart Component code comments - updated badges in Readme - updated yml file push to master from main
