+++ example/package.json
@@ -10,8 +10,7 @@
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
   "dependencies": {
-    "@shopify/react-native-skia": "^0.1.211",
-    "d3": "^7.8.5",
+    "@shopify/react-native-skia": "^0.1.228",
     "react": "18.1.0",
     "react-native": "0.70.0",
     "react-native-graph-kit": "../"
@@ -21,7 +20,6 @@
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
     "@tsconfig/react-native": "^2.0.2",
-    "@types/d3": "^7.4.1",
     "@types/jest": "^26.0.23",
     "@types/react-native": "^0.69.6",
     "@types/react-test-renderer": "^18.0.0",
