+++ package.json
@@ -1,10 +1,9 @@
 {
   "name": "react-native-graph-kit",
-  "version": "0.0.1",
-  "description": "React native line and bar graphs through skia",
+  "version": "1.0.0",
+  "description": "Personalized graphs featuring customizable options for React Native app",
   "main": "lib/index",
   "types": "lib/index.d.ts",
-  "contributors": [],
   "author": "Simform Solutions",
   "repository": {
     "type": "git",
@@ -15,7 +14,18 @@
     "react",
     "react-native",
     "typescript",
-    "rn"
+    "rn",
+    "chart",
+    "graph",
+    "visualization",
+    "line-chart",
+    "bar-chart",
+    "react-native-skia",
+    "data-visualization",
+    "open-source",
+    "javascript",
+    "skia",
+    "react-native-library"
   ],
   "license": "MIT",
   "files": [
@@ -32,10 +42,10 @@
     "example": "yarn --cwd example"
   },
   "peerDependencies": {
-    "react": "*",
-    "react-native": "*",
     "@shopify/react-native-skia": "^0.1.211",
-    "d3": "^7.8.5"
+    "d3-scale": "^4.0.2",
+    "react": "*",
+    "react-native": "*"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
@@ -44,25 +54,25 @@
     "@commitlint/config-conventional": "^16.0.0",
     "@react-native-community/eslint-config": "^3.0.1",
     "@testing-library/react-native": "^9.0.0",
-    "babel-jest": "^27.4.6",
-    "typescript": "4.7.4",
-    "eslint-plugin-simple-import-sort": "^7.0.0",
-    "jest": "^27.4.7",
-    "husky": "^8.0.1",
-    "lint-staged": "^11.1.2",
-    "prettier": "^2.7.1",
-    "react": "18.1.0",
-    "react-native": "0.70.0",
     "@tsconfig/react-native": "^2.0.2",
+    "@types/d3-scale": "^4.0.8",
     "@types/jest": "^27.4.0",
     "@types/react-native": "^0.69.5",
     "@types/react-test-renderer": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^5.29.0",
     "@typescript-eslint/parser": "^5.29.0",
+    "babel-jest": "^27.4.6",
     "eslint": "^7.32.0",
+    "eslint-plugin-simple-import-sort": "^7.0.0",
+    "husky": "^8.0.1",
+    "jest": "^27.4.7",
+    "lint-staged": "^11.1.2",
     "metro-react-native-babel-preset": "^0.70.3",
+    "prettier": "^2.7.1",
+    "react": "18.1.0",
+    "react-native": "0.70.0",
     "react-test-renderer": "18.0.0",
-    "@types/d3": "^7.4.1"
+    "typescript": "4.7.4"
   },
   "husky": {
     "hooks": {
@@ -101,7 +111,7 @@
     ]
   },
   "dependencies": {
-    "@shopify/react-native-skia": "^0.1.211",
-    "d3": "^7.8.5"
+    "@shopify/react-native-skia": "^0.1.228",
+    "d3-scale": "^4.0.2"
   }
 }
