+++ src/components/bar-chart/components/yaxis-label/YAxisLabelsPropsTypes.ts
@@ -1,5 +1,5 @@
 import type { SkFont } from '@shopify/react-native-skia';
-import type { ScaleLinear } from 'd3';
+import type { ScaleLinear } from 'd3-scale';
 
 export interface YAxisLabelsPropsType {
   font: SkFont;
