+++ src/components/line-chart/components/y-axis-labels/YAxisLabelsTypes.ts
@@ -1,5 +1,5 @@
 import type { Color, SkFont } from '@shopify/react-native-skia';
-import type { ScaleLinear } from 'd3';
+import type { ScaleLinear } from 'd3-scale';
 
 export type YAxisLabelsPropsType = {
   yScale: ScaleLinear<number, number>;
